refactor: move Extra type into pattern_command.py, remove dead actions/{pattern_action,destinations_aware_action,group_action}.py
Extra was only re-exported from actions/pattern_action.py; PatternCommand
(the Typer-era replacement) already re-implemented the same logic and
just imported the type from the old module. All 8 live importers now
import Extra from typer.base_commands.pattern_command instead.

destinations_aware_action.py and group_action.py had zero importers
left after this branch already removed the 18 legacy action files that
used them — removed as dead code, no separate freeze branch needed
since they're already preserved on all 18 refactor/migrate-*-to-typer
branches (forked before this branch's removals).
