+++ src/justin/typer/upload_command.py
@@ -24,7 +24,7 @@ from pyvko.pyvko_main import Pyvko
 from typer import Typer, Argument
 
 from justin.actions.mixins import EventUtils
-from justin.actions.pattern_action import Extra
+from justin.typer.base_commands.pattern_command import Extra
 from justin.cms_2.storage.sqlite.sqlite_entries import Person
 from justin.shared.helpers.parts import folder_tree_parts
 from justin.shared.metafiles.metafile import PostMetafile, PostStatus, GroupMetafile, PersonMetafile, CommentMetafile, \
